bluewaters craycc fixes

+++ main-mpi.c
@@ -38,7 +38,7 @@ int main(int argc, char * argv[]) {
         return 1;
     }
 
-    size_t NUMITEMS = atoi(argv[1]);
+    size_t NUMITEMS = atoll(argv[1]);
 
     size_t mysize = (ThisTask + 1 ) * NUMITEMS / NTask - ThisTask * NUMITEMS / NTask;
     int * mydata = malloc(mysize * sizeof(int));
